requirements.txt
streamlit>=1.42.0
google-api-python-client>=2.100.0
supabase>=2.0.0
plotly>=5.18.0
pandas>=2.1.0
python-dotenv>=1.0.0
isodate>=0.6.1
anthropic>=0.30.0
